requirements.txt
beanie==1.29.0
fastapi==0.115.11
fastapi_users==14.0.1
fastapi_users_db_beanie==4.0.0
mangum==0.19.0
motor==3.7.0
pydantic==2.10.6
python-dotenv==1.0.1
